+++ R/pairwise_comparisons.R
@@ -692,13 +692,14 @@ pairwise_test_bin <- function(x,
 
 
 
-#' Correlation Pairwise Testing
+#' Correlation Testing for Multiple Groups/Measures
 #'
-#' Takes a continuous variable and a grouping variable to calculate the pairwise
-#' Spearman, Pearson, or Kendall correlation estimate and p-value between two variables.
+#' Takes a continuous variable and a categorical variable, and calculate the
+#' pairwise Spearman, Pearson, or Kendall correlation estimate and p-value
+#' between the categorical variable levels.
 #'
 #' @param x numeric vector (can include NA values)
-#' @param group categorical vector which contains the group levels to compare
+#' @param group categorical vector which contains the levels to compare
 #' @param id vector which contains the id information
 #' @param method character string indicating which correlation coefficient
 #'   is to be used for the test ("pearson" (default), "kendall", or "spearman").
@@ -725,12 +726,11 @@ pairwise_test_bin <- function(x,
 #' @return Returns a data frame of all possible pairwise correlations
 #' with group sizes greater than or equal to the minimum number of values
 #' in group, as set by `n_distinct_value`:
-#' * `Comparison` - Comparisons made
-#' * `DistinctValues` - number of distinct points
+#' * `Correlation` - Comparisons made
 #' * `NPairs` - number of non-missing pairs considered
+#' * `Ties` - are ties present in either variable
 #' * `CorrEst` - correlation estimates
 #' * `CorrTest` - correlation test p value
-#' * Unpasted columns if `keep_vars = TRUE`
 #' @details
 #'
 #' The p value is calculated using the [cor_test] function (see documentation
@@ -753,7 +753,7 @@ pairwise_test_bin <- function(x,
 #'   bb = c(NA,NA,NA,NA,NA,1:5)
 #' )
 #' data_in_long <- tidyr::pivot_longer(data_in, -id)
-#' pairwise_test_cor(x = data_in_long$value,
+#' cor_test_pairs(x = data_in_long$value,
 #'                   group = data_in_long$name,
 #'                   id = data_in_long$id,
 #'                   method = 'spearman')
@@ -770,7 +770,7 @@ pairwise_test_bin <- function(x,
 #' filter(visitno != 0) %>%
 #' group_by(group, visitno) %>%
 #'  summarize(
-#'    pairwise_test_cor(x = magnitude, group = antigen, id = pubID,
+#'    cor_test_pairs(x = magnitude, group = antigen, id = pubID,
 #'    method = 'spearman', n_distinct_value = 3, digits = 1, verbose = TRUE),
 #'    .groups = 'drop'
 #'           )
@@ -778,7 +778,7 @@ pairwise_test_bin <- function(x,
 #' @export
 
 
-pairwise_test_cor <- function(x,
+cor_test_pairs <- function(x,
                               group,
                               id,
                               method = c('spearman', 'pearson', 'kendall'),
@@ -844,10 +844,22 @@ pairwise_test_cor <- function(x,
       data_here <- stats::na.omit(merge(i_data, j_data, by = 'id'))
       N_points <- nrow(data_here)
 
-      comparison_here <- paste0(i_group, ' vs. ', j_group)
+      comparison_here <- paste0(i_group, ' and ', j_group)
 
-      distinct_vals <- paste0(length(unique(data_here$x)),
-                              ' vs. ',
+      dups_x <- any(duplicated(data_here$x))
+      dups_y <- any(duplicated(data_here$y))
+      if (dups_x & dups_y) {
+        ties = 'ties in both'
+      } else if (dups_x) {
+        ties = paste0('ties in ', i_group)
+      } else if (dups_y) {
+        ties = paste0('ties in ', j_group)
+      } else {
+        ties = "no ties"
+      }
+
+        paste0(length(unique(data_here$x)),
+                              ' and ',
                               length(unique(data_here$y)))
 
       if (N_points > 0) {
@@ -883,9 +895,9 @@ pairwise_test_cor <- function(x,
       }
 
       results_list[[length(results_list) + 1]] <-
-        data.frame(Comparison = comparison_here,
+        data.frame(Correlation = comparison_here,
                    NPoints = N_points,
-                   DistinctValues = distinct_vals,
+                   Ties = ties,
                    CorrEst = rho,
                    CorrTest = mag_p,
                    stringsAsFactors = FALSE)