+++ tests/testthat/test_pairwise_comparisons.R
@@ -781,14 +781,14 @@ test_that("error checking", {
     drop_na()
 
   expect_error(
-    object =   pairwise_test_cor(x = x,
+    object =   cor_test_pairs(x = x,
                                  group = group,
                                  id = 1:10),
     regexp = '"x", "group", and "id" must be same length'
   )
 
   expect_error(
-    object =   pairwise_test_cor(x = x,
+    object =   cor_test_pairs(x = x,
                                  group = group,
                                  id = id,
                                  n_distinct_value = 1),
@@ -796,24 +796,24 @@ test_that("error checking", {
   )
 
   expect_error(
-    object =   pairwise_test_cor(x = rep(1:2, 26),
+    object =   cor_test_pairs(x = rep(1:2, 26),
                                  group = group,
                                  id = id),
     regexp = 'All groups have less than 3 distinct values'
   )
 
   expect_error(
-    object =   pairwise_test_cor(x = c(rep(1:2, 25), 3, 4),
+    object =   cor_test_pairs(x = c(rep(1:2, 25), 3, 4),
                                  group = group,
                                  id = id),
     regexp = 'Only one group has >=3 distinct values, so no testing possible'
   )
 
   expect_message(
-    object = pairwise_test_cor(x = c(NA,NA,NA,x[-(1:3)],1,2,3),
+    object = cor_test_pairs(x = c(NA,NA,NA,x[-(1:3)],1,2,3),
                                    group = c(group,rep('c',3)),
                                    id = c(id,1:3), verbose = TRUE),
-    regexp = 'No non-missing data points when considering a vs. c'
+    regexp = 'No non-missing data points when considering a and c'
   )
 
 
@@ -823,8 +823,8 @@ test_that("error checking", {
 
 
 
-# test pairwise_test_cor Using paste_tbl_grp and cor_test in testing since these functions are testing elsewhere
-test_that("pairwise_test_cor testing two groups", {
+# test cor_test_pairs Using paste_tbl_grp and cor_test in testing since these functions are testing elsewhere
+test_that("cor_test_pairs testing two groups", {
   library(tidyr)
   library(dplyr)
   set.seed(243542534)
@@ -839,11 +839,15 @@ test_that("pairwise_test_cor testing two groups", {
 
 
   testing_results <- data.frame(
-    Comparison = 'a vs. b',
+    Correlation = 'a and b',
     NPoints = nrow(test_data_wide),
-    DistinctValues = paste0(test_data_wide$a %>% na.omit %>% length,
-                            ' vs. ',
-                            test_data_wide$b %>% na.omit %>% length),
+    Ties = case_when(
+      any(duplicated(test_data_wide$a)) &
+        any(duplicated(test_data_wide$b)) ~ 'ties in both',
+      any(duplicated(test_data_wide$a)) ~ 'ties in a',
+      any(duplicated(test_data_wide$b)) ~ 'ties in b',
+      TRUE ~ 'no ties'
+    ),
     CorrEst = stats::cor(x = test_data_wide$a,
                        y = test_data_wide$b,
                        method = 'spearman', use = 'pairwise.complete.obs') %>%
@@ -854,7 +858,7 @@ test_that("pairwise_test_cor testing two groups", {
     stringsAsFactors = FALSE
   )
 
-  expect_equal(object = pairwise_test_cor(x = test_data$x,
+  expect_equal(object = cor_test_pairs(x = test_data$x,
                                              group = test_data$group,
                                              id = test_data$id,
                                              method = 'spearman',
@@ -865,7 +869,7 @@ test_that("pairwise_test_cor testing two groups", {
                expected = testing_results)
 
   expect_message(
-    object = pairwise_test_cor(x = test_data$x,
+    object = cor_test_pairs(x = test_data$x,
                                   group = test_data$group,
                                   id = test_data$id,
                                   method = 'spearman',
@@ -884,7 +888,7 @@ test_that("pairwise_test_cor testing two groups", {
                                         use = 'pairwise.complete.obs') %>%
     round_away_0(digits = 5, trailing_zeros = TRUE)
 
-  expect_equal(object = pairwise_test_cor(x = test_data$x,
+  expect_equal(object = cor_test_pairs(x = test_data$x,
                                              group = test_data$group,
                                              id = test_data$id,
                                              method = 'spearman',
@@ -898,8 +902,8 @@ test_that("pairwise_test_cor testing two groups", {
 
 
 
-# test pairwise_test_cor Using paste_tbl_grp and cor_test in testing since these functions are testing elsewhere
-test_that("pairwise_test_cor testing multiple groups", {
+# test cor_test_pairs Using paste_tbl_grp and cor_test in testing since these functions are testing elsewhere
+test_that("cor_test_pairs testing multiple groups", {
   library(tidyr)
   library(purrr)
 
@@ -914,13 +918,19 @@ test_that("pairwise_test_cor testing multiple groups", {
       drop_na()
 
     testing_results <- data.frame(
-      Comparison = paste0(levels(test_data$group)[1],
-                          ' vs. ',
+      Correlation = paste0(levels(test_data$group)[1],
+                          ' and ',
                           levels(test_data$group)[2]),
       NPoints = nrow(test_data_wide),
-      DistinctValues = paste0(test_data_wide[,2, drop = TRUE] %>% length,
-                              ' vs. ',
-                              test_data_wide[,3, drop = TRUE] %>% length),
+      Ties = case_when(
+        any(duplicated(test_data_wide[,2, drop = TRUE])) &
+          any(duplicated(test_data_wide[,3, drop = TRUE])) ~ 'ties in both',
+        any(duplicated(test_data_wide[,2, drop = TRUE])) ~
+          paste0('ties in ', levels(test_data$group)[1]),
+        any(duplicated(test_data_wide[,3, drop = TRUE])) ~
+          paste0('ties in ', levels(test_data$group)[2]),
+        TRUE ~ 'no ties'
+      ),
       CorrEst = ifelse(sum(!test_data_wide[,2, drop = TRUE] %>% duplicated) >= 3 &
                          sum(!test_data_wide[,3, drop = TRUE] %>% duplicated) >= 3,
         stats::cor(x = test_data_wide[,2, drop = TRUE],
@@ -965,7 +975,7 @@ test_that("pairwise_test_cor testing multiple groups", {
 
   group_testing_dt <- testData_BAMA %>%
     group_by(group, visit) %>%
-    group_modify(~pairwise_test_cor(x = .x$magnitude,
+    group_modify(~cor_test_pairs(x = .x$magnitude,
                                     group = .x$antigen,
                                     id = .x$pubID,
                                     method = 'spearman',
@@ -981,7 +991,7 @@ test_that("pairwise_test_cor testing multiple groups", {
   expect_message(
     object = testData_BAMA %>%
       group_by(group, visit) %>%
-      group_modify(~pairwise_test_cor(x = .x$magnitude,
+      group_modify(~cor_test_pairs(x = .x$magnitude,
                                    group = .x$antigen,
                                    id = .x$pubID,
                                    method = 'spearman',
@@ -989,7 +999,7 @@ test_that("pairwise_test_cor testing multiple groups", {
                                    digits = 3,
                                    trailing_zeros = TRUE,
                                    verbose = TRUE)),
-    regexp = 'Not enough distinct values for at least one group when considering Con S gp140 CFI vs. gp70_C.1086C V1/V2/293F'
+    regexp = 'Not enough distinct values for at least one group when considering Con S gp140 CFI and gp70_C.1086C V1/V2/293F'
   )
 
 })